shopmanagementapp/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
Django==4.2.7
django-filter==23.4
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==20.1.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.2
pillow==10.4.0
psycopg==3.1.13
pycparser==2.21
PyJWT==2.9.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==73.0.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0